refactor: place font reading and processing in a package
Read the URW font from an io.Reader and build up the font line by line.

Sort of make use of an interface so that you can add an image font
reader later, but it is not really used right now.

Add stringer methods to the different components of the font for easy
printing of the font.