feat: custom viewport that keeps selected char at the centre
The custom viewport allows for highlighting the selected pixel in the
selected char, but also highlight the corresponding pixel in other chars
so you can align edits that you make to other chars.

The viewport also keeps the selected Char at the center so you have a
consistent view of it.

Navigate to other chars using the keyboard or mouse and navigate to
other pixels within the selected char using the keyboard or mouse.